gcc updates
